Kernel: Use the templated copy_to/from_user() in more places
These ensure that the "to" and "from" pointers have the same type,
and also that we copy the correct number of bytes.

+++ Kernel/Thread.h
@@ -365,7 +365,7 @@ public:
     FPUState& fpu_state() { return *m_fpu_state; }
 
     void set_default_signal_dispositions();
-    void push_value_on_stack(u32);
+    void push_value_on_stack(uintptr_t);
 
     u32 make_userspace_stack_for_main_thread(Vector<String> arguments, Vector<String> environment);
